fix(codegen): P0 — bare-dot templates prove roles, accumulators and every value path per role
The single-dot matmul templates (inline simdgroup, K-loop, strided scalar, delegated
make_matmul_kernel) took several things on faith and were GPU-verified silently wrong on
clean ba21e9d: the output written into an unrelated INPUT pointer (positional roles), fused
accumulators dropped, second stores / atomics / residual base offsets dropped, operand and
result casts dropped, K-loop pointer streams bound by DECLARATION ORDER (err 345), a
non-canonical K-loop mask and a non-BLOCK_K pointer advance ignored, the tutorial's 1-D
grid split dispatched as 2-D (60% of C unwritten), and an fp16 dot result exposed as the
raw f32 accumulator. Reviewed adversarially with GPT (packets 065-096); approved on these
exact file hashes (091/095):

    476b122629b068ad  triton_msl/codegen/_lowerer_detection.py
    ef38fd08f9b725f0  triton_msl/codegen/generic_lowerer.py
    d7d78b7452c8907b  triton_msl/codegen/_lowerer_templates.py
    f7d587795b756c80  tests/test_simple_dot_roles_and_acc.py
    ccadfd8c24e7ecd3  tests/test_cpp_backend.py

Four dependency-ordered parts, landed as one tree (the parts are interleaved inside the
value-path predicate and the templates; the sign-off attaches to this exact tree):

1. Roles through loops (W6). _trace_ptr_source(iter_inits=) + _scf_iter_inits cross an
   scf.for iter-arg to its init (opt-in; 14 other call sites unchanged);
   _resolve_dot_ptr_roles drops its declaration-order fallback (9 callers audited:
   infer_dot_strides and the batched-base-offset guard no longer guess positionally).

2. Per-role value-path proof (W1-W5, W7, 081/083). _dot_template_value_paths proves, for
   A/B: load -> layout ops only -> dot with equal element types, the load reading the
   traced address (through iter-args), K-loop masks only in the template's own clip
   (both spellings, iteration-counted loops, 1-D compares broadcast by expand_dims), a
   literal-zero `other`; for C: dot/loop result -> convert_layout -> at most the ONE cast
   the template replays -> the single store, with an f16/bf16 dot result admitted only
   when the store IS its result conversion (sole-observation census; extension to f32
   refuses); for the K-loop: accumulator <-> iter-arg <-> yield through layout ops only,
   pointer streams advanced by exactly BLOCK_K * the traced K stride or indexed by
   `k + range`; int8 -> i32 admitted by EXACTNESS (single tile, K <= 1023, i32 output).
   infer_dot_strides(with_index=) exports each axis's 1-D index (opt-in). Casts leave the
   op allowlist; only the proven output cast is admitted, by id. Positive literal-zero
   accumulator proof replaces the negative bias test (K-loop twin scoped to the
   accumulator iter-arg). Every kernel arg is declared at its own buffer slot; the
   delegated make_matmul_kernel route refuses unless roles sit at 0/1/2.

3. Grid mapping (W8). Tile indices must be [program_id(axis) * BLOCK +] make_range
   [+ k] with rows on axis 0 / cols on axis 1, or the tutorial's 1-D split whose divisor
   is provably cdiv(<the M extent the template clips with>, BLOCK_M); the K-loop and
   strided-scalar templates emit the PROVEN mapping (_pid_map_lines: 1d | 2d, no
   fallback), returned by the predicate rather than kept as lowerer state.

4. Tests. tests/test_simple_dot_roles_and_acc.py: 58 pins (GPU witnesses that must refuse
   or match the IR-ordered oracle, positives that must stay on the template and be exact,
   rule-9 lowering-boundary pins); on clean ba21e9d 39 fail / 18 pass / 1 skip. Two C++-path
   tests use native fp16 buffers (their in-kernel fp32->fp16 cast was the W1 class, hidden
   by a 0.5 tolerance).

Gates on this tree: pins 58/0; GPT holdouts 6/6; focused slice 1859/0; randomized project
gate 2566/0/16 (= 2508 + 58); fixed-order upstream ratchet 158/5366/3818, identical to the
ba21e9d baseline (dot3d 135 + dot_multidim 20 + masked_load 3). Perf: no material
regression in the sampled kernel-throughput paths (evidence dir perf-2026-09-02).

+++ tests/test_cpp_backend.py
@@ -441,16 +441,20 @@ def test_cpp_dot_32x32():
             off_k = tl.arange(0, BLOCK_K)
             a = tl.load(a_ptr + off_m[:, None] * K + off_k[None, :])
             b = tl.load(b_ptr + off_k[:, None] * N + off_n[None, :])
-            c = tl.dot(a.to(tl.float16), b.to(tl.float16), acc=tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32))
+            c = tl.dot(a, b, acc=tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32))
             tl.store(c_ptr + off_m[:, None] * N + off_n[None, :], c)
 
+        # Native fp16 buffers (packet 079): an in-kernel ``.to(tl.float16)`` of fp32
+        # loads is a value transform the Python matmul templates never replayed
+        # (silently staged the raw fp32 buffer) and now refuse — and the Python
+        # lowering always runs first as the C++ path's fallback source.
         M = N = K = 32
-        a = torch.randn(M, K)
-        b = torch.randn(K, N)
+        a = torch.randn(M, K).half()
+        b = torch.randn(K, N).half()
         c = torch.zeros(M, N)
         matmul_kernel[(1,)](a, b, c, M=M, N=N, K=K, BLOCK_M=M, BLOCK_N=N, BLOCK_K=K)
 
-        expected = a @ b
+        expected = a.float() @ b.float()
         max_err = (c - expected).abs().max().item()
         # f16 tolerance (relaxed)
         assert max_err < 0.5, f"32x32 matmul: max error {max_err}"
@@ -530,17 +534,18 @@ def test_cpp_dot_k_loop():
                 off_k = k_off + tl.arange(0, BLOCK_K)
                 a = tl.load(a_ptr + off_m[:, None] * K + off_k[None, :])
                 b = tl.load(b_ptr + off_k[:, None] * N + off_n[None, :])
-                acc = tl.dot(a.to(tl.float16), b.to(tl.float16), acc)
+                acc = tl.dot(a, b, acc)
             tl.store(c_ptr + off_m[:, None] * N + off_n[None, :], acc)
 
+        # Native fp16 buffers (packet 079) — see test_cpp_dot_32x32.
         M = N = 16
         K = 32
-        a = torch.randn(M, K)
-        b = torch.randn(K, N)
+        a = torch.randn(M, K).half()
+        b = torch.randn(K, N).half()
         c = torch.zeros(M, N)
         matmul_k_loop[(1,)](a, b, c, M=M, N=N, K=K, BLOCK_M=M, BLOCK_N=N, BLOCK_K=16)
 
-        expected = a @ b
+        expected = a.float() @ b.float()
         max_err = (c - expected).abs().max().item()
         assert max_err < 0.5, f"k-loop matmul: max error {max_err}"
     finally:
